pyproject.toml
[project]
name = "match-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "google-genai>=1.52.0",
    "python-dotenv>=1.2.1",
    "yt-dlp[default]>=2025.11.12",
]
